Bump typescript to v4
Fix type issue in benchmark-utils


Bump @types/eventsource to support Typescrip v4

Remove un-used toHex() function


Cast pubkey to Uint8Array in lodestar-validator

Cast querystring.parse to fastify querystringParser


networkInterfaces maybe undefined


Declare Promise resolve type void


Remove unused import ArrayLike


Revert deleting toHex()

+++ packages/lodestar-validator/src/slashingProtection/block/blockBySlotRepository.ts
@@ -54,7 +54,10 @@ export class BlockBySlotRepository {
   }
 
   private encodeKey(pubkey: BLSPubkey, slot: Slot): Buffer {
-    return encodeKey(this.bucket, Buffer.concat([Buffer.from(pubkey), intToBytes(BigInt(slot), uintLen, "be")]));
+    return encodeKey(
+      this.bucket,
+      Buffer.concat([Buffer.from(pubkey as Uint8Array), intToBytes(BigInt(slot), uintLen, "be")])
+    );
   }
 
   private decodeKey(key: Buffer): {pubkey: BLSPubkey; slot: Slot} {
